refactors the p4c driver code to allow more flexibility for backends
Different backends may carry special configurations, flags, and
intermediate passes that were not captured in the current design.

To that end, we define a Driver class that each backend instantiates
and configures its commands, setup and cleanup steps, as well as the
options to pass to commands.

For now, the design simply replicates the commands -- in fact for BMV2
and EBPF we do not take much advantage of the additional flexibility.