Fix resubmit in bmv2 backend
Just like for learn lists, field list ids (which are used for resubmit)
need to be non-zero, since the id is actually used as a flag to indicate
whether or not resubmit was invoked.

Given that the "resubmit_flag" v1model metadata field is supposed to
hold a field list id, its bitwidth was increased from 1 to 32.