Fix handling of p4runtime options in p4c driver
The --p4runtime-file option provided by the user was ignored by the
driver (for bmv2), which was adding its own --p4runtime-file option when
invoking the compiler. For some reason, the bnv2 backend was the only
one to work this way. With this commit, the driver no longer generates
its own option by default and instead honors the one provided by the
user.

Fixes #1221