Fix a bug with the backtracking framework (#1425)
When a backtrack trigger of type `t` is thrown from a nested pass manager and when it is meant to be
caught in the top-level pass manager, the trigger type currently does not recognize type `t`. This
PR ensures that the trigger type is not lost when a backtrack exception is rethrown by the nested
pass manager.