Add getOutputStream method to ErrorReporter
I would like to write some GTests in which I change the output stream to
be a stringstream so that the unit test can verify that the error
message is the expected one. Adding the getOutputStream method enables
me to "save" the existing output stream then "restore" it when I am
done.